# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "CCS API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/hc-cases/search": {
            "get": {
                "tags": [
                    "High Court"
                ],
                "summary": "Search High Court cases",
                "description": "Searches the hc_case_full_details table using optional filters.",
                "operationId": "f486bb07901f4c5bb24c7998db92afe2",
                "parameters": [
                    {
                        "name": "pet_name",
                        "in": "query",
                        "description": "Filter by petitioner name (partial match)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "Amit"
                        }
                    },
                    {
                        "name": "pet_adv",
                        "in": "query",
                        "description": "Filter by petitioner advocate (partial match)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "Adv. Sharma"
                        }
                    },
                    {
                        "name": "res_name",
                        "in": "query",
                        "description": "Filter by respondent name (partial match)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "State of J&K"
                        }
                    },
                    {
                        "name": "res_adv",
                        "in": "query",
                        "description": "Filter by respondent advocate (partial match)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "Adv. Khan"
                        }
                    },
                    {
                        "name": "sub_category",
                        "in": "query",
                        "description": "Filter by sub category (partial match)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "Civil"
                        }
                    },
                    {
                        "name": "est_code",
                        "in": "query",
                        "description": "Filter by establishment code (exact match)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "JKOJ"
                        }
                    },
                    {
                        "name": "case_type",
                        "in": "query",
                        "description": "Filter by case type id (exact match)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 171
                        }
                    },
                    {
                        "name": "reg_year",
                        "in": "query",
                        "description": "Filter by registration year (exact match)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 2023
                        }
                    },
                    {
                        "name": "reg_no",
                        "in": "query",
                        "description": "Filter by registration number (exact match)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "328"
                        }
                    },
                    {
                        "name": "cino",
                        "in": "query",
                        "description": "Filter by CINO (exact match)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "JKHC010002862013"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Results per page (default 20, max 100)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "maximum": 100,
                            "minimum": 1,
                            "example": 20
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number (default 1)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "minimum": 1,
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful search",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        },
                                        "meta": {
                                            "properties": {
                                                "current_page": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "per_page": {
                                                    "type": "integer",
                                                    "example": 20
                                                },
                                                "total": {
                                                    "type": "integer",
                                                    "example": 200
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}